# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaksi Pekerjaan API",
        "description": "API untuk aplikasi **Transaksi Pekerjaan** — pencatatan arus kas per kegiatan\nbeserta perhitungan bagi hasil.\n\n### Alur autentikasi\n1. `POST /api/auth/login` dengan email + password → dapat `access_token`.\n2. Kirim `Authorization: Bearer <access_token>` untuk semua endpoint lain.\n3. Opsional: `POST /api/auth/biometric/enable` (butuh password lagi) →\n   dapat `biometric_token` yang disimpan aplikasi di Android Keystore.\n   Login berikutnya cukup sidik jari lewat `POST /api/auth/biometric/login`.\n\n### Aturan angka\nSemua nilai uang adalah **rupiah bulat** (integer, tanpa desimal).\nPersentase dikirim sebagai angka desimal biasa, mis. `1.75` untuk 1,75%.",
        "contact": {
            "email": "bayuapriansah10@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/laporan/ringkasan": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Ringkasan dashboard",
                "description": "Angka untuk dashboard: jumlah kegiatan per status, akumulasi pagu & profit, posisi kas, tren kas bulanan, dan kegiatan dengan profit tertinggi. Semua bisa dibatasi periode.",
                "operationId": "laporanRingkasan",
                "parameters": [
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "bulan_tren",
                        "in": "query",
                        "description": "Jumlah bulan pada grafik tren kas",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 6,
                            "maximum": 24,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/laporan/rekap-kegiatan": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Rekap tabel kegiatan",
                "description": "Tabel rekap semua kegiatan dengan kolom seperti di Excel \"Taksasi Pekerjaan\", plus baris TOTAL. Dipakai layar Laporan di mobile.",
                "operationId": "laporanRekapKegiatan",
                "parameters": [
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "berjalan",
                                "selesai",
                                "batal"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/laporan/kegiatan/{id}/pdf": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Cetak PDF taksasi kegiatan",
                "description": "Mencetak laporan satu kegiatan ke PDF: identitas kegiatan, rincian taksasi baris-per-baris beserta persentasenya, dan daftar arus kas. Respons berupa berkas PDF (application/pdf).",
                "operationId": "laporanKegiatanPdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "download",
                        "in": "query",
                        "description": "1 = paksa unduh (attachment), 0 = tampilkan inline",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "sertakan_kas",
                        "in": "query",
                        "description": "1 = sertakan daftar arus kas",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berkas PDF",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kegiatan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/laporan/rekap-kegiatan/pdf": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Cetak PDF rekap kegiatan",
                "description": "Mencetak tabel rekap seluruh kegiatan ke PDF (orientasi landscape) beserta baris TOTAL.",
                "operationId": "laporanRekapPdf",
                "parameters": [
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "download",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berkas PDF",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/laporan/arus-kas/pdf": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Cetak PDF buku kas",
                "description": "Mencetak buku kas ke PDF: daftar transaksi berurutan tanggal dengan kolom debet/kredit dan saldo berjalan. Bisa dibatasi per kegiatan dan per periode.",
                "operationId": "laporanArusKasPdf",
                "parameters": [
                    {
                        "name": "kegiatan_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "download",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berkas PDF",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan access_token dari endpoint login (tanpa kata \"Bearer\").",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}